requirements.txt
accelerate==1.10.1
acres==0.5.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
argcomplete==1.10.3
asgiref==3.9.1
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
babel==2.17.0
bcc==0.18.0
beautifulsoup4==4.8.2
bitsandbytes==0.47.0
blinker==1.4
Brlapi==0.8.3
Brotli==1.1.0
certifi==2025.7.14
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.4.3
chrome-gnome-shell==0.0.0
ci-info==0.3.0
click==8.0.3
colorama==0.4.4
comm==0.2.3
command-not-found==0.3
compressed-rtf==1.0.7
configobj==5.0.9
configparser==7.2.0
cryptography==45.0.5
cssselect2==0.8.0
cupshelpers==1.0
dbus-python==1.2.18
debugpy==1.8.16
decorator==5.2.1
defer==1.0.6
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==5.2.4
django-allauth==65.9.0
django-ckeditor==6.7.3
django-js-asset==3.1.2
django-parler==2.3
django-widget-tweaks==1.5.0
docx2txt==0.9
docxcompose==1.4.0
docxtpl==0.20.1
dotenv==0.9.9
ebcdic==1.1.1
etelemetry==0.3.1
exceptiongroup==1.3.0
executing==2.2.1
extract-msg==0.28.7
Faker==37.4.2
filelock==3.18.0
fonttools==4.59.2
frontend==0.0.3
frozenlist==1.7.0
fsspec==2025.9.0
geoip2==5.1.0
git-filter-repo==2.47.0
greenlet==3.2.3
groq==0.30.0
gunicorn==23.0.0
gyp==0.1
h11==0.16.0
h2==4.2.0
hf-xet==1.1.9
hidpidaemon==18.4.6
hpack==4.1.0
httpcore==1.0.9
httplib2==0.20.2
httpx==0.28.1
huggingface-hub==0.34.4
hyperframe==6.1.0
idna==3.3
IMAPClient==2.1.0
importlib-metadata==4.6.4
importlib_resources==6.5.2
ipykernel==6.30.1
ipython==8.37.0
isodate==0.7.2
itsdangerous==2.2.0
jedi==0.19.2
jeepney==0.7.1
Jinja2==3.1.6
jupyter_client==8.6.3
jupyter_core==5.8.1
kaitaistruct==0.10
kazam==1.4.5
kernelstub==3.1.4
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
looseversion==1.3.0
louis==3.20.0
lxml==6.0.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
matplotlib-inline==0.1.7
maxminddb==2.7.0
more-itertools==8.10.0
mpmath==1.3.0
multidict==6.6.3
nest-asyncio==1.6.0
netaddr==0.8.0
netifaces==0.11.0
networkx==3.4.2
nibabel==5.3.2
nipype==1.10.0
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
oauthlib==3.2.0
olefile==0.47
opencv-python==4.12.0.88
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.1
parso==0.8.5
pathlib==1.0.1
pdf2image==1.17.0
pdfminer.six==20191110
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.4.0
playwright==1.53.0
pop-transition==1.1.2
prompt_toolkit==3.0.52
propcache==0.3.2
protobuf==3.12.4
prov==2.1.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
puremagic==1.30
pyasn1==0.6.1
pycairo==1.20.1
pycparser==2.22
pycryptodome==3.23.0
pycups==2.0.1
pydantic==2.11.7
pydantic_core==2.33.2
pydbus==0.6.0
pydot==4.0.1
pydyf==0.11.0
pyee==13.0.0
Pygments==2.19.2
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyMuPDF==1.26.3
PyNaCl==1.5.0
pyOpenSSL==25.1.0
pyparsing==3.2.3
pyphen==0.17.2
pyRFC3339==1.1
PySocks==1.7.1
pytesseract==0.3.13
python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-decouple==3.8
python-docx==1.2.0
python-dotenv==1.1.1
python-gnupg==0.4.8
python-pptx==0.6.23
python-xlib==0.29
pytz==2022.1
pyxdg==0.27
pyxnat==1.6.3
PyYAML==5.4.1
pyzmq==27.1.0
rdflib==7.1.4
regex==2025.9.1
repolib==2.2.1
repoman==1.4.0
reportlab==4.4.3
requests==2.32.5
safetensors==0.6.2
scipy==1.15.3
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
selenium==4.35.0
selenium-wire==5.1.0
simplejson==3.20.1
six==1.12.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
SpeechRecognition==3.8.1
sqlparse==0.5.3
stack-data==0.6.3
starlette==0.47.1
stripe==12.5.0
sympy==1.14.0
systemd-python==234
textract==1.6.5
tinycss2==1.4.0
tinyhtml5==2.0.0
tokenizers==0.22.0
torch==2.8.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
traits==7.0.2
transformers==4.56.1
trio==0.30.0
trio-websocket==0.12.2
triton==3.4.0
TwoCaptcha==0.0.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
undetected-chromedriver==3.5.5
urllib3==2.5.0
user-agents==2.2.0
uvicorn==0.35.0
wadllib==1.3.6
wcwidth==0.2.13
weasyprint==66.0
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
whitenoise==6.9.0
wsproto==1.2.0
xdg==5
xkit==0.0.0
xlrd==1.2.0
xlsxwriter==3.2.9
yarl==1.20.1
zipp==1.0.0
zopfli==0.2.3.post1
zstandard==0.23.0
